% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.excellent.R
\name{count.excellent}
\alias{count.excellent}
\title{Counts number of excellent fit indices}
\usage{
count.excellent(
  indices,
  rmsea.cut = 0.05,
  srmr.cut = 0.05,
  nnfi.cut = 0.95,
  cfi.cut = 0.95
)
}
\arguments{
\item{indices}{A vector of fit indices from lavaan.}
}
\value{
The number of fit indices that are excellent.
}
\description{
Counts number of excellent fit indices
}
\keyword{internal}
