% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.gimme.R
\name{multiverse.gimme}
\alias{multiverse.gimme}
\title{Multiverse GIMME}
\usage{
multiverse.gimme(
  data,
  groupcutoffs = 0.75,
  subcutoffs = 0.51,
  rmsea.cuts = 0.05,
  srmr.cuts = 0.05,
  nnfi.cuts = 0.95,
  cfi.cuts = 0.95,
  n.excellent = 2,
  n.cores = 1,
  prune_output = TRUE,
  save_output = FALSE,
  save_dir = NULL,
  ...
)
}
\arguments{
\item{data}{The data for fitting the models.}

\item{groupcutoffs}{The cutoff values for group-level fit indices. Should be a numeric value or a vector of numeric values.}

\item{subcutoffs}{The cutoff values for subgroup-level fit indices. Should be a numeric value or a vector of numeric values.}

\item{rmsea.cuts}{The cutoff values for RMSEA (Root Mean Square Error of Approximation) fit indices.}

\item{srmr.cuts}{The cutoff values for SRMR (Standardized Root Mean Residual) fit indices.}

\item{nnfi.cuts}{The cutoff values for NNFI (Non-Normed Fit Index) fit indices.}

\item{cfi.cuts}{The cutoff values for CFI (Comparative Fit Index) fit indices.}

\item{n.excellent}{The number of excellent fitting indices to require for model selection.}

\item{n.cores}{The number of cores to use for parallel fitting. Set to 1 for nonparallel fitting.}

\item{prune_output}{Logical; if TRUE, the model data will be removed from the output list for reduced memory usage.}

\item{save_output}{Logical; if TRUE, every iteration will be saved as .RDS}

\item{save_dir}{Directory where results should be saved if \code{save_output = TRUE}. Will be created if it does not yet exist.}

\item{...}{Additional arguments to be passed to \code{\link[mvgimme::gimme]{mvgimme::gimme}} function.}
}
\value{
A list containing the results of the multiverse analysis for different parameter combinations. Each element of the list corresponds to one specification, and the conditions used for fitting are attached as a data frame to each result.
}
\description{
This function is a wrapper around the `gimme` function. It allows fitting a multiverse of models with different specifications in parallel.
}
